Fixed Bug 10414: "Thin multisegmented custom lines incorrectly rendered to PDF"
git-svn-id: svn://scribus.net/branches/Version135/Scribus@17047 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pdflib_core.cpp
@@ -1842,13 +1842,13 @@ bool PDFLibCore::PDF_TemplatePage(const Page* pag, bool )
 					QVector<double>::iterator it;
 					for ( it = ite->DashValues.begin(); it != ite->DashValues.end(); ++it )
 					{
-						int da = static_cast<int>(*it);
+						double da = *it;
 						// #8758: Custom dotted lines don't export properly to pdf
 						// Null values have to be exported if line end != flat
 						if ((da != 0) || (ite->lineEnd() != Qt::FlatCap))
 							PutPage(QString::number(da)+" ");
 					}
-					PutPage("] "+QString::number(static_cast<int>(ite->DashOffset))+" d\n");
+					PutPage("] "+QString::number(ite->DashOffset)+" d\n");
 				}
 				else
 					PutPage("["+getDashString(ite->PLineArt, ite->lineWidth())+"] 0 d\n");
@@ -3092,13 +3092,13 @@ QString PDFLibCore::PDF_ProcessTableItem(PageItem* ite, const Page* pag)
 		QVector<double>::iterator it;
 		for ( it = ite->DashValues.begin(); it != ite->DashValues.end(); ++it )
 		{
-			int da = static_cast<int>(*it);
+			double da = *it;
 			// #8758: Custom dotted lines don't export properly to pdf
 			// Null values have to be exported if line end != flat
 			if ((da != 0) || (ite->lineEnd() != Qt::FlatCap))
 				tmp += QString::number(da)+" ";
 		}
-		tmp += "] "+QString::number(static_cast<int>(ite->DashOffset))+" d\n";
+		tmp += "] "+QString::number(ite->DashOffset)+" d\n";
 	}
 	else
 		tmp += "["+getDashString(ite->PLineArt, ite->lineWidth())+"] 0 d\n";
@@ -3230,13 +3230,13 @@ bool PDFLibCore::PDF_ProcessItem(QString& output, PageItem* ite, const Page* pag
 		QVector<double>::iterator it;
 		for ( it = ite->DashValues.begin(); it != ite->DashValues.end(); ++it )
 		{
-			int da = static_cast<int>(*it);
+			double da = *it;
 			// #8758: Custom dotted lines don't export properly to pdf
 			// Null values have to be exported if line end != flat
 			if ((da != 0) || (ite->lineEnd() != Qt::FlatCap))
 				tmp += QString::number(da)+" ";
 		}
-		tmp += "] "+QString::number(static_cast<int>(ite->DashOffset))+" d\n";
+		tmp += "] "+QString::number(ite->DashOffset)+" d\n";
 	}
 	else
 		tmp += "["+getDashString(ite->PLineArt, ite->lineWidth())+"] 0 d\n";
@@ -4312,13 +4312,13 @@ bool PDFLibCore::setTextCh(PageItem *ite, uint PNr, double x,  double y, uint d,
 				QVector<double>::iterator it;
 				for ( it = embedded->DashValues.begin(); it != embedded->DashValues.end(); ++it )
 				{
-					int da = static_cast<int>(*it);
+					double da = *it;
 					// #8758: Custom dotted lines don't export properly to pdf
 					// Null values have to be exported if line end != flat
 					if ((da != 0) || (embedded->lineEnd() != Qt::FlatCap))
 						tmp2 += QString::number(da)+" ";
 				}
-				tmp2 += "] "+QString::number(static_cast<int>(embedded->DashOffset))+" d\n";
+				tmp2 += "] "+QString::number(embedded->DashOffset)+" d\n";
 			}
 			else
 				tmp2 += "["+getDashString(embedded->PLineArt, embedded->lineWidth())+"] 0 d\n";
@@ -5106,13 +5106,13 @@ bool PDFLibCore::PDF_Gradient(QString& output, PageItem *currItem)
 				QVector<double>::iterator it;
 				for ( it = item->DashValues.begin(); it != item->DashValues.end(); ++it )
 				{
-					int da = static_cast<int>(*it);
+					double da = *it;
 					// #8758: Custom dotted lines don't export properly to pdf
 					// Null values have to be exported if line end != flat
 					if ((da != 0) || (item->lineEnd() != Qt::FlatCap))
 						tmp2 += QString::number(da)+" ";
 				}
-				tmp2 += "] "+QString::number(static_cast<int>(item->DashOffset))+" d\n";
+				tmp2 += "] "+QString::number(item->DashOffset)+" d\n";
 			}
 			else
 				tmp2 += "["+getDashString(item->PLineArt, item->lineWidth())+"] 0 d\n";
